Set new cli main as new main.py
